User can share other users' tweets

+++ app/models/User.scala
@@ -3,7 +3,7 @@ package models
 case class User(
                  username: String,
                  password: String,
-                 var sharedTweets: List[Int], // list of tweets shared by this user
+                 var sharedTweets: List[Long], // list of tweets shared by this user
                  var followedUsers: List[String] = List()
                )
